final_competition/requirements.txt
nltk==3.6.2
numpy==1.18.5
regex==2021.4.4
scikit-learn==0.24.2
scipy==1.6.3
torchaudio==0.8.1
torchtext==0.4.0
